requirements.txt
Flask~=1.1
requests~=2.22
pycryptodome
